move get-vacant-pair-list into domain-service
