+++ src/config/participant-router.ts
@@ -10,7 +10,7 @@ import { ParticipantFactory } from "../domain/participant/participant-factory"
 import { PairFactory } from "../domain/pair/pair-factory"
 import { GeneratePairName } from "../domain/pair/domain-service/generate-pair-name"
 import { JoinPair } from "../domain/pair/join-pair"
-import { GetVacantPairList } from "../domain/pair/get-vacant-pair-list"
+import { GetVacantPairList } from "../domain/pair/domain-service/get-vacant-pair-list"
 import { GetParentTeam } from "../domain/pair/domain-service/get-parent-team"
 
 // repository
