+++ src/domain/pair/domain-service/get-vacant-pair-list.ts
@@ -1,6 +1,6 @@
-import { DomainService } from "../shared/domainService"
-import { Pair } from "./pair"
-import { IPairRepository } from "./interface/pair-repository"
+import { DomainService } from "../../shared/domainService"
+import { Pair } from "../pair"
+import { IPairRepository } from "../interface/pair-repository"
 
 export class GetVacantPairList extends DomainService<"get-vacant-pair-list"> {
   constructor(private readonly pairRepository: IPairRepository) {
